requirements.txt
google-genai>=0.3.0
SpeechRecognition>=3.10.0
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pyaudio>=0.2.13
psutil>=5.9.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
vosk>=0.3.45
numpy>=1.24.0
torch>=2.0.0
torchaudio>=2.0.0
speechbrain>=0.5.15
pywin32>=306
mss>=9.0.0
edge-tts>=6.1.9
pygame>=2.5.0
openai>=1.1.0
groq>=0.3.0
